requirements.txt
anyio==4.8.0
APScheduler==3.11.0
argcomplete==3.1.4
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
blinker==1.7.0
boto3==1.34.46
botocore==1.34.46
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==24.4.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
cryptography==41.0.7
dbus-python==1.3.2
distro==1.9.0
distro-info==1.7+build1
et_xmlfile==2.0.0
export-module==0.0.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jalali_core==1.0.0
jdatetime==5.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
netaddr==0.8.0
netifaces==0.11.0
numpy==2.2.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.0
pandas==2.2.3
persian-tools==0.0.11
pexpect==4.9.0
pipx==1.4.3
platformdirs==4.2.0
psutil==5.9.8
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
python-magic==0.4.27
python-telegram-bot==22.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
s3transfer==0.10.1
schedule==1.2.2
service-identity==24.1.0
setuptools==68.1.2
six==1.16.0
sniffio==1.3.1
sos==4.8.2
ssh-import-id==5.11
supervisor==4.2.5
systemd-python==235
Twisted==24.3.0
typing_extensions==4.10.0
tzdata==2025.1
tzlocal==5.3
ubuntu-pro-client==8001
unattended-upgrades==0.1
urllib3==2.0.7
userpath==1.9.1
wadllib==1.3.6
wheel==0.42.0
XlsxWriter==3.2.3
xyzservices==0.0.0
zope.interface==6.1
